Reusable composable specs
